# Optimize rule/marginalrule dispatch