--- a/docs/src/custom/custom-functional-form.md
+++ b/docs/src/custom/custom-functional-form.md
@@ -14,6 +14,7 @@ q(x) = f\left(\frac{\overrightarrow{\mu}(x)\overleftarrow{\mu}(x)}{\int \overrig
 
 ```@docs 
 AbstractFormConstraint
+CompositeFormConstraint
 ```
  
 ### Form check strategy