--- a/src/nodes/probit.jl
+++ b/src/nodes/probit.jl
@@ -18,8 +18,9 @@ default_meta(::Type{Probit}) = ProbitMeta(32)
 
 default_functional_dependencies_pipeline(::Type{<:Probit}) = RequireMessageFunctionalDependencies((2,), (vague(NormalMeanPrecision),))
 
-default_interface_local_constraint(::Type{<:Probit}, edge::Val{:in})  = MomentMatching()
-default_interface_local_constraint(::Type{<:Probit}, edge::Val{:out}) = Marginalisation()
+interface_default_local_constraint(::Type{<:Probit}, edge::Symbol) = interface_default_local_constraint(Probit, Val(edge))
+interface_default_local_constraint(::Type{<:Probit}, edge::Val{:in})  = MomentMatching()
+interface_default_local_constraint(::Type{<:Probit}, edge::Val{:out}) = Marginalisation()
 
 @average_energy Probit (q_out::Union{PointMass, Bernoulli}, q_in::UnivariateNormalDistributionsFamily, meta::ProbitMeta) = begin