… docs/src/lib/message.md
… docs/src/lib/nodes/flow.md
… docs/src/lib/nodes/nodes.md
… docs/src/lib/rules/rules.md
… src/rules/gamma_mixture/switch.jl
… src/rules/normal_mixture/switch.jl
… src/score/node.jl
… src/score/score.jl
… test/nodes/test_bifm_helper.jl
… test/nodes/test_gamma_inverse.jl
… test/nodes/test_mv_normal_mean_covariance.jl
… test/nodes/test_mv_normal_mean_precision.jl
… test/nodes/test_mv_normal_mean_scale_precision.jl
… test/nodes/test_normal_mean_precision.jl
… test/nodes/test_normal_mean_variance.jl
… test/nodes/test_normal_mixture.jl
… test/nodes/test_poisson.jl
… test/nodes/test_probit.jl
… test/nodes/test_uniform.jl
… test/nodes/test_wishart_inverse.jl
… test/runtests.jl
… test/test_rule.jl