@bvdmitri 2023-03-23
Thanks for fixing the docs @bartvanerp . Are you okay with the remaining changes? Or did you plan to change text in other places?